+++ src/main/java/org/team4/common/logger/Logger.java
@@ -3,7 +3,6 @@ package org.team4.common.logger;
 import javafx.application.Platform;
 import org.team4.App;
 import org.team4.common.Settings;
-import org.team4.exceptionClass.LogException;
 import org.team4.dashboard.DashboardController;
 
 import java.io.FileWriter;
@@ -13,7 +12,6 @@ import java.util.Date;
 public class Logger {
     public static String fileName = "output.jsonl";
 
-
     /**
      * Creates a log, saves it to a file and display it in the output box
      * @param message of the log
@@ -21,7 +19,7 @@ public class Logger {
      * @param date current date/time
      * @param user current user
      */
-    public static boolean log(String message, String level, Date date, String user){
+    public static boolean log(String message, String level, Date date, String user) {
         Log newLog = new Log(message, level, date, user);
         writeToLogFile(newLog);
         Platform.runLater(
@@ -37,7 +35,7 @@ public class Logger {
      * Generate a log level info
      * @param message
      */
-    public static boolean info(String message){
+    public static boolean info(String message) {
         String user = Settings.currentUser;
         Date date = Settings.simulationTime.getDate();
         String level = "info";
@@ -74,15 +72,6 @@ public class Logger {
         try {
             FileWriter myWriter = new FileWriter(fileName, true);
             String logStr = log.toJson().toString();
-            //check whether the information that user entered is correct
-            try {
-                if(logStr==null||logStr.equals("")){
-                    throw new LogException();
-                }
-
-            }catch (LogException e){
-                e.printStackTrace();
-            }
             myWriter.write(logStr + "\n");
             myWriter.close();
         } catch (IOException e) {
